+++ src/components/Content.tsx
@@ -1,3 +1,22 @@
+import { useMovie } from '../movieContext'
+import { MovieCard } from "./MovieCard"
+
 export function Content() {
-  // Complete aqui
+  const { selectedGenre, movies } = useMovie()
+
+  return(
+    <div className="container">
+      <header>
+        <span className="category">Categoria:<span> {selectedGenre.title}</span></span>
+      </header>
+
+      <main>
+        <div className="movies-list">
+          {movies.map(movie => (
+            <MovieCard key ={movie.imdbID} title={movie.Title} poster={movie.Poster} runtime={movie.Runtime} rating={movie.Ratings[0].Value} />
+          ))}
+        </div>
+      </main>
+    </div>
+  )
 }
