+++ src/styles/movie-card.scss
