+++ src/styles/sidebar.scss
